+++ src/main/java/org/cancogenvirusseq/singularity/api/ApiController.java
@@ -28,9 +28,9 @@ import lombok.extern.slf4j.Slf4j;
 import org.cancogenvirusseq.singularity.api.model.EntityListResponse;
 import org.cancogenvirusseq.singularity.api.model.ErrorResponse;
 import org.cancogenvirusseq.singularity.components.base.DownloadObjectById;
+import org.cancogenvirusseq.singularity.components.pipelines.Contributors;
 import org.cancogenvirusseq.singularity.exceptions.ArchiveNotFoundException;
 import org.cancogenvirusseq.singularity.exceptions.BaseException;
-import org.cancogenvirusseq.singularity.pipelines.Contributors;
 import org.cancogenvirusseq.singularity.repository.ArchivesRepo;
 import org.cancogenvirusseq.singularity.repository.model.Archive;
 import org.cancogenvirusseq.singularity.repository.model.ArchiveType;
