+++ src/main/java/org/cancogenvirusseq/singularity/components/base/DownloadObjectById.java
@@ -10,10 +10,10 @@ import java.util.function.UnaryOperator;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.cancogenvirusseq.singularity.components.model.AwsSdkFluxResponse;
+import org.cancogenvirusseq.singularity.components.utils.FluxResponseProvider;
 import org.cancogenvirusseq.singularity.config.s3Client.S3ClientProperties;
 import org.cancogenvirusseq.singularity.exceptions.DownloadFailedException;
 import org.cancogenvirusseq.singularity.repository.model.Archive;
-import org.cancogenvirusseq.singularity.utils.FluxResponseProvider;
 import org.springframework.stereotype.Component;
 import reactor.core.publisher.Mono;
 import software.amazon.awssdk.http.SdkHttpResponse;
