+++ src/main/java/org/cancogenvirusseq/singularity/components/pipelines/Contributors.java
@@ -16,7 +16,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.cancogenvirusseq.singularity.pipelines;
+package org.cancogenvirusseq.singularity.components.pipelines;
 
 import java.util.Arrays;
 import java.util.Set;
